random to ordered questions
In the previous code, questions were choosen from a list randomly, which often caused repetition of questions (You could be asked for the capital fo the same country twice and not be asked about other countries ). But now instead, I shuffle the list and loop through each country in the list. So you are always asked about all the countries in your chosen continent.